requirements.txt
discord.py
python-dotenv
requests
bs4
setuptools
sqlite3
fake-useragent